refac: 파일 로드 시점 동기화 문제로 useImageMetadata와 useImageFile 다시 합침
